runtime: fix plan9 warning.
I have no idea what this code is for, but it pretty
clearly needs to be uint64, not uint32.

LGTM=aram
R=0intro, aram
CC=golang-codereviews
https://golang.org/cl/84410043

+++ src/pkg/runtime/os_plan9_amd64.c
@@ -95,7 +95,7 @@ runtime·sighandler(void *v, int8 *note, G *gp)
 		if(ureg->ip != 0) {
 			sp = (uintptr*)ureg->sp;
 			*--sp = ureg->ip;
-			ureg->sp = (uint32)sp;
+			ureg->sp = (uint64)sp;
 		}
 		ureg->ip = (uintptr)runtime·sigpanic;
 		return NCONT;
